scripts/requirements.txt
python-frontmatter
persiantools
requests
beautifulsoup4